requirements.txt
djangorestframework-jwt
djangorestframework
asgiref==3.2.7
beautifulsoup4==4.8.2
certifi==2020.4.5
chardet==3.0.4
Django==3.0.4
idna==2.9
joblib==0.14.1
numpy==1.18.2
pandas==1.0.3
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.8
